+++ src/config/config.ts
@@ -77,11 +77,15 @@ export interface ConfigPropertyMetaInput {
 export class Config extends ConfigFile<ConfigFile.Options> {
   /**
    * Username associated with the default dev hub org.
+   *
+   * @deprecated Replaced by OrgConfigProperties.TARGET_DEV_HUB in v3 {@link https://github.com/forcedotcom/sfdx-core/blob/v3/MIGRATING_V2-V3.md#config}
    */
   public static readonly DEFAULT_DEV_HUB_USERNAME: string = 'defaultdevhubusername';
 
   /**
    * Username associate with the default org.
+   *
+   * @deprecated Replaced by OrgConfigProperties.TARGET_ORG in v3 {@link https://github.com/forcedotcom/sfdx-core/blob/v3/MIGRATING_V2-V3.md#config}
    */
   public static readonly DEFAULT_USERNAME: string = 'defaultusername';
 
@@ -104,6 +108,11 @@ export class Config extends ConfigFile<ConfigFile.Options> {
    * Disables telemetry reporting
    */
   public static readonly DISABLE_TELEMETRY = 'disableTelemetry';
+
+  /**
+   * Custom templates repo or local location.
+   */
+  public static readonly CUSTOM_ORG_METADATA_TEMPLATES = 'customOrgMetadataTemplates';
 
   /**
    * allows users to override the 10,000 result query limit
@@ -133,6 +142,16 @@ export class Config extends ConfigFile<ConfigFile.Options> {
         validator: (value) => value == null || (isString(value) && sfdc.validateApiVersion(value)),
         get failedMessage() {
           return Config.messages?.getMessage('InvalidApiVersion');
+        },
+      },
+    },
+    {
+      key: Config.CUSTOM_ORG_METADATA_TEMPLATES,
+      input: {
+        // If a value is provided validate it otherwise no value is unset.
+        validator: (value) => value == null || isString(value),
+        get failedMessage() {
+          return Config.messages?.getMessage('InvalidCustomOrgMetadataTemplates');
         },
       },
     },